finished process of transactions (Connecting FE and BE)

+++ PSPBackend/src/main/java/com/psp/PSPBackend/dto/AuthRequest.java
@@ -16,6 +16,6 @@ public class AuthRequest {
     private String merchantId;
     private String merchantPassword;
     private Double amount;
-    private Integer merchantOrderId;
+    private Long merchantOrderId;
     private LocalDateTime merchantTimeStamp;
 }
